Changed unisigned int types to uint16_t

+++ TimeDelay.c
@@ -10,7 +10,7 @@
 #include "TimeDelay.h"
 
 //method to start a timer for time_ms milliseconds
-void Delay_time(unsigned int time_ms){
+void Delay_time(uint16_t time_ms){
     T2CONbits.T32 = 1;//allows Timer2 and Timer3 to form a 32-bit timer
     T2CONbits.TSIDL = 0;//allows timer to continue while cpu is in idle mode.
